Add `Rescaler` component (#275)

+++ src/bin/update_snapshots/main.rs
@@ -20,11 +20,18 @@ use crate::{
 fn main() {
     println!("Updating snapshots:");
 
-    let test: Vec<_> = snapshot_tests()
-        .into_iter()
-        .map(TestCaseInstance::new)
-        .collect();
-    for test in test.iter() {
+    let tests: Vec<_> = snapshot_tests();
+    let has_only_flag = tests.iter().any(|t| t.only);
+    let tests: Vec<_> = if has_only_flag {
+        tests
+            .into_iter()
+            .filter(|t| t.only)
+            .map(TestCaseInstance::new)
+            .collect()
+    } else {
+        tests.into_iter().map(TestCaseInstance::new).collect()
+    };
+    for test in tests.iter() {
         for pts in &test.case.timestamps {
             let (snapshots, Err(_)) = test.test_snapshots_for_pts(*pts) else {
                 println!("PASS: \"{}\" (pts: {}ms)", test.case.name, pts.as_millis());
@@ -63,16 +70,16 @@ fn main() {
             }
         }
     }
-
-    // Check for unused snapshots
-    let snapshot_paths = test
-        .iter()
-        .flat_map(TestCaseInstance::snapshot_paths)
-        .collect::<HashSet<_>>();
-    for path in find_unused_snapshots(&snapshot_paths, snapshots_path()) {
-        println!("Removed unused snapshot {path:?}");
-        fs::remove_file(path).unwrap();
+    if !has_only_flag {
+        // Check for unused snapshots
+        let snapshot_paths = tests
+            .iter()
+            .flat_map(TestCaseInstance::snapshot_paths)
+            .collect::<HashSet<_>>();
+        for path in find_unused_snapshots(&snapshot_paths, snapshots_path()) {
+            println!("Removed unused snapshot {path:?}");
+            fs::remove_file(path).unwrap();
+        }
     }
-
     println!("Update finished");
 }
